perf: removed unnecessary content pre-rendering from server render process
For subsequent loads, the browser just needs the state to render the
page itself, and the Wasm is already instantiated, so there's no reason
to generate the content on the server as well, which slows page loads
down and increases the amount needing to be transferred over the
network. The gains from this commit are minimal, but present.