feat: added page state store caching, preloading, and memory management (#204)
* feat: added eviction to the pss

This should fix any memory blowouts (not leaks, just the storage of
every single page's state can get a bit heavy), which occasionally
occurred in development.

* chore: clarified a comment

* feat: created facility to support storage of document metadata in pss

* feat: prevented network requests for cached head/state

After a page is fetched as a subsequent load, it won't be fetched again
until the PSS fills up! This doesn't work with initial loads, since the
head is pre-interpolated (and I don't want to increase the bundle size
by doubling it up in a variable; reading from the HTML is unreliable
since JS will likely have already modified it).

* feat: added preloading infrastructure

This has involved making `RenderCtx` hold much more data in the browser,
like the error pages and render context, though this should make routing
much lighter.

* fix: fixed longstanding clones in app route system

Apparently, it's perfectly valid to pass the Sycamore scope through to
the Sycamore router, which means we an access everything we need from
the render context! (I reckon there'll be a performance improvement to
moving the render context into a dedicated system though, beyond
Sycamore's context.)

* refactor: removed unnecessary if clause in fetching example

I think this led to some confusion the other day, so it's clarified now.
Just that we don't need `G::IS_BROWSER` if we're target-gating as well.

* feat: created user-facing preload system

This includes a new `core/preload` example.

* chore: applied #212 fix to new `preload` example

* feat: added initially loaded page caching

This is achieved through an extra `<meta>` delimiter that denotes the
end of the `<head>`, which should be pretty reliable at getting what the
user intended.

* feat: added feature to control initial page caching

Advanced `<head>` manipulations *could* in rare cases lead to bugs, so
the user can turn this off if necessary, and it's documented in the FAQ section.