feat: added `#[engine_only_fn]`
This is the simplest macro in history, and is now used everywhere to
clearly mark functions that are engine-side-only. This is a huge
improvement over the opacity of `#[build_state]` and co., which did this
implicitly. Also, this macro is simple enough to avoid easily if a user wants to.