feat: made axum integration work with turbines
It and the core now compile (not for the browser-side yet though). This
also brings nearly all the server-side logic into the core, which should
make authoring future integrations a trivial matter of just extracting
the correct data.

This currently uses a deliberate memory leak to get a `&'static Turbine`
to avoid the use of an `Arc`. So far, I see no issues with this, since
it only happens once, but this API may change to an `Arc`.