feat: made view functions fallible
This makes error handling *much* cleaner, and removes a very large
number of panics that hypothetically wouldn't have been executed on the
browser-side, but the invariants were not rigorously guaranteed. They
now are, and errors are gracefully returned for handling.

This involves *substantial* import restructuring, and a greater
dependence on `perseus::prelude` for end users.