feat: made templates with state work without macros!
This involves some serious GAT/HRTB magic, and I don't fully
understand *why* this compiles yet, but it does. Tradeoffs are that the
state type does have to be manually specified, and a `BoundedScope<'a,
'b>` is needed in the user's function signature, which isn't great, but
it's alright, and that can be explained to them with documentation.

I want to try to remove the requirement to specify the state type...