feat: restructed reference reactive state
This changes the use of `&'a RcSignal<T>` to use `create_ref()` on
the entire intermediate state, allowing more ergonomic reference
passing, such as passing the entire state into a closure to call
a user-defined method.

Although this is a very substantial conceptual change, the code changes
involved are actually fairly small.

Additionally, this fixes type inference on `.template_with_state()`.