feat: added incremental generation with capsules examples
Note that this was the moment I realized incremental generation with
widgets can be used with reactivity to reactively command the server to
generate widgets in advance. That is unreasomably cool, and would allow
something like generating the widgets that back product pages on an
e-commerce website through the flipping search bar. The pieces are
coming together...