feat: remove localized global state (#268)
* feat: removed localized global state

This completely axes the feature, and returns global state to a
one-per-app system.

BREAKING CHANGE: removed localized global state (global state functions
should no longer take a locale)

* fix: transmitted global state in locale redirection pages

This fixes #267, but only when global state is unlocalized.

* fix: fixed broken examples

* fix: fixed broken demo example

* chore(deps): updated dependencies extensively

Not strictly relevant to this PR, but important nonetheless. Note that
this includes a progression of the `minify-html-onepass` crate version,
which *could* cause problems, so that should be reverted to v0.10.1 if
there are issues.